Gemini_version/requirements.txt
web3==6.15.1
pandas==2.2.0
python-dotenv==1.0.0
numpy>=1.24.0
TA-Lib>=0.4.24
scikit-learn>=1.2.0 